refactor: clean departments
